Split parser into its own module
